requirements.txt
Django==1.4.1
MySQL-python==1.2.4b2
PIL==1.1.7
South==0.7.6
django-tastypie==0.9.15
lxml==3.1beta1
simplejson==3.3.1

